% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFPWM.R
\name{createFPWM}
\alias{createFPWM}
\title{A function to generate an FPWM class object.}
\usage{
createFPWM(
  mainTF = NULL,
  partners = NULL,
  cell = NULL,
  mainTF_MMID = NULL,
  partners_MMID = NULL,
  forkPosition = NULL,
  probabilityMatrix = FALSE,
  scaleFrequencyCounts = FALSE,
  flipMatrix = FALSE
)
}
\arguments{
\item{mainTF}{[character] character with the name of the main TF.}

\item{partners}{[character list or vector] List or character vector with the names of the partner TFs.}

\item{cell}{[character] with the name of the main TF.}

\item{mainTF_MMID}{[character] with the name of the main TF.}

\item{partners_MMID}{[character vector] with the name of the main TF.}

\item{forkPosition}{[numeric] defines the postion in the matrix where the motif will be forked.}

\item{probabilityMatrix}{[logical] whether the function should return a frequency matrix or probability matrix (Default FALSE).}

\item{scaleFrequencyCounts}{[logical] whether the count matrix should have equal rowSums across all the rows (Default FALSE).}

\item{flipMatrix}{[logical] whether to apply reverse complement in case the core motif is after the forkPosition (Default FALSE).}
}
\value{
returns a FPWM class object that can be used to plot or write in transfact format.
}
\description{
This function creates the FPWM object and performs the fork in the motif.
}
\examples{
fpwm <- createFPWM(mainTF ="CEBPB", partners = c("ATF4","ATF7","JUND","CEBPD"), cell = "K562", forkPosition = 5)
}
