% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniCofactorReport.R
\name{miniCofactorReport}
\alias{miniCofactorReport}
\title{Produces a miniCofactorReport}
\usage{
miniCofactorReport(
  TF,
  cell,
  cobinding_threshold = 0.05,
  Methylation = TRUE,
  includeMotifOnly = TRUE,
  height = 12,
  width = 7,
  pdfName = NULL
)
}
\arguments{
\item{TF}{[character] Main TF of interest.}

\item{cell}{[character] Cell of interest.}

\item{cobinding_threshold}{[numeric] Only the co-factors with co-binding percentages more than this threshold value will be reported. By default the threshold is 0.05.}

\item{Methylation}{[logical] TRUE to retrieve cytosine methylation information.}

\item{includeMotifOnly}{[logical] TRUE if you wish to include only peaks that contain the known binding motif.}

\item{height}{[numeric] Height in inch for the plot.}

\item{width}{[numeric] Width in inch for the plot.}

\item{pdfName}{[character] Name of the pdf to be saved.}
}
\value{
A PDF file with the cofactorReport.
}
\description{
This function allows you to get a PDF report of top cofactors along with DNA methylation information for a motif of a TF.
}
\examples{
miniCofactorReport(TF = "CEBPB",cell = "K562")
}
\keyword{cofactorReport}
